Refactoring the layout.
I modified the layout in tk because I want to present the preview in full width. (eventually also the lesson contents too) Moving the nav to the topbar to always have it available in the same place.
I really dislike UI that moves arund when the layout changes.
Some other layout changes like centering the videos, and adding / removing paddings here and there.