feat: tooltips (#23)
* feat: WIP tooltips

* refactor: removing complicated logic and leveraging library's API to prevent style injection.

* feat: polishing work on the tooltips (#24)

Had to bring in styles from TK or the portaled tooltpi wouldn't work. Needed to portal the tooltip so it would show on top of the rest of the UI, even on narrower screens.

* Adding styles for the links, they weren't working

* fix: tooltips for mobile

* Before git renaming the file

* Renaming to get back the tracking

* Fixing the freaking tooltips once more.

They worked on local, but not once deployed.
So I went back to the version that did work for just placing the tooltip with the anchor, and then added the 'portaling' on top.

* fix: mobile styles

* fixing nav background style